second dongle
added `-d 1` to the rtl_sdr command in launch_978 to force dump978 to use the second dongle. Note that the preferred solution would be to have separate dongles dedicated to a specific frequency, but in my case I'm using two stock antennas near each other in the same window. It doesn't matter which frequency each dongle monitors for traffic.